% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conciliacao.R
\name{conciliacao}
\alias{conciliacao}
\title{Compara as duas tabelas, a do banco e a do admp.}
\usage{
conciliacao(dados_admp, dados_banco)
}
\arguments{
\item{dados_admp}{equivale aos dados exportados pelo programa ADMP em csv.}

\item{dados_banco}{equivale aos dados exportados pelo seu banco também em csv. Atenção a esse dado porque cada banco exporta os dados de uma determinada maneira.}
}
\value{
A comparação entre \code{dados_admp} e \code{dados_banco}.
}
\description{
Se algo estiver errado, é importante conferir a movimentação de cada dotação naquele mês.
sdasd
}
\examples{
conciliacao(1, 1)
conciliacao(10, 1)
}
